simulation/requirements.txt
numpy
scipy
pyserial
matplotlib
pyquaternion